+++ bonsai-py/scripts/README.md
@@ -1,16 +1,16 @@
 # bonsai-py scripts
 
-Developer scripts for working on the `bonsai_py` extension module.
+Developer scripts for working on the `bonsai_bt` extension module.
 
 ## Scripts
 
 | Script | Purpose |
 |---|---|
-| [regen-stubs.sh](regen-stubs.sh) | Regenerates `python/bonsai_py/__init__.pyi` from the `#[gen_stub_*]` annotations on the Rust side. Run after editing any annotated `#[pyclass]` / `#[pyfunction]` / `#[pymethods]`. Also runs automatically via the `regen-stubs` pre-commit hook and is enforced in CI. |
+| [regen-stubs.sh](regen-stubs.sh) | Regenerates `python/bonsai_bt/__init__.pyi` from the `#[gen_stub_*]` annotations on the Rust side. Run after editing any annotated `#[pyclass]` / `#[pyfunction]` / `#[pymethods]`. Also runs automatically via the `regen-stubs` pre-commit hook and is enforced in CI. |
 
 ## Prerequisites
 
-A Python venv with the `bonsai_py` extension built in. See [../README.md](../README.md#installation-dev) for the one-time setup (`python -m venv .venv`, activate, `pip install maturin`, `maturin develop --release`).
+A Python venv with the `bonsai_bt` extension built in. See [../README.md](../README.md#installation-dev) for the one-time setup (`python -m venv .venv`, activate, `pip install maturin`, `maturin develop --release`).
 
 ## Running
