+++ bonsai-py/tests/test_telemetry.py
@@ -5,7 +5,7 @@ from typing import Any
 
 import pytest
 
-import bonsai_py as bt
+import bonsai_bt as bt
 
 
 class TestWithTelemetry:
